requirements.txt
aws-cdk.core==1.170.0
aws-cdk.aws_ec2==1.170.0
aws-cdk.aws_elasticloadbalancingv2==1.170.0
